Changed return structure of the CDP scan
The scan now returns a tuple for each MAC address instead of a single IP address.
The first field of the tuple is a list of all the devices IP addresses.
The second field is the preferred IP address which should be routable from the scanner.

The order of preference is currently IPv4 addresses and link-local IPv6 addresses.